Add files via upload
Given a positive integer n, the following rules will always create a sequence that ends with 1, called the hailstone sequence: 

- If n is even, divide it by two
- If n is odd, multiply it by 3 and add 1 (i.e. 3n +1)
- Continue until n is 1

Write a program that reads an integer as input and prints the hailstone sequence starting with the integer entered. Format the output so that ten integers, each separated by a tab character (\t), are printed per line. End the last line of output with a new line (endl).