Add files via upload
Write a program that counts the number of times each letter of the alphabet appears in a given text. You will be given a text, Hamlet's famous "to be or not to be" speech, as a string - as well as a pre-filled vector of each character of the alphabet. You will need to loop through each character and count how many times it appears in the text. Then you will need to determine which letter appears the most.

The output of your program should show each character's count in the text. It should also show the character that appears the most times, and the number of times it appears.