Add files via upload
You will be given vectors pre-filled with data about squirrel behavior in your program template. Write a program that analyzes this data, and prints out the following:

Total number of squirrels observed: #
Number of squirrels eating: #
Number of squirrels indifferent: #
Number of squirrels both eating and indifferent: #

In this example, the # is just a placeholder for the real integer that you will find in your program. You will need to count the number of total squirrels, the total number of squirrels that have been marked as "eating," and the total number of squirrels that have been labeled by observers as "indifferent" (in other words, not paying attention to the observer at all). You will also need to count the number of squirrels that have been marked as both "eating" and "indifferent."

Note: the vectors given in the code are parallel vectors, meaning that each item in the vector matches the item in the other vector, at the same index. In other words, squirrelsEating.at(0) and squirrelsIndifferent.at(0) both refer to observations about the same squirrel.