Add files via upload
Write a program that stores a list of positive integers from input into a vector and outputs the Nth number from the end of the vector. A negative integer indicates the end of the input and is not stored in the vector. Convert the negative integer read at the end to positive and use as N. Output the negative integer read at the end if the size of the vector is smaller than N. Assume the vector will always contain fewer than 20 integers.